better memory footprint - 27270901 in .9 seconds

+++ src/atom.rs
@@ -1,11 +1,11 @@
 #[derive(Debug)]
 pub struct Nucleus {
-  pub protons: i8,
-  pub neutrons: i8,
+  pub protons: i16,
+  pub neutrons: i16,
 }
 
 #[derive(Debug)]
 pub struct Atom {
-  pub electrons: i8,
+  pub electrons: i16,
   pub nucleus: Nucleus,
 }
